+++ spec/spec_helper.rb
@@ -9,6 +9,11 @@ RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
 
+  config.before do
+    TelegramSupportBot.reset_adapter!
+    TelegramSupportBot.reset_state_store! if TelegramSupportBot.respond_to?(:reset_state_store!)
+  end
+
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!